% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirstfun.R
\name{mtfirstfun}
\alias{mtfirstfun}
\title{My first R function}
\usage{
mtfirstfun(x)
}
\arguments{
\item{x}{A vector of quantitative data}
}
\value{
A vector of squared componets
}
\description{
My first R function
}
\examples{
y <- 1:10; myfirstfun(y)
}
