% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpPlot.R
\name{SimpPlot}
\alias{SimpPlot}
\title{Simple Scatter Plot with set Equation}
\usage{
SimpPlot(x)
}
\arguments{
\item{x}{An x value from 0 ot 50 that will return a plot according to the equation}
}
\value{
A plot of x vs y
}
\description{
Simple Scatter Plot with set Equation
}
\examples{
x <- 0:50; SimpPlot(x)
}
